added helpers to one module
