+++ mop/__init__.py
@@ -11,9 +11,8 @@ __all__ = ['cemba',
            'io',
            'loom_utils',
            'plot',
-           'plot_helpers',
+           'helpers',
            'recipes',
-           'recipes_helpers',
            'qc',
            'smooth',
            'snmcseq',
